backend/auth-service/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
httpx==0.28.1
pydantic==2.12.4
pydantic-settings
python-multipart==0.0.20
python-dotenv==1.2.1
